For computing the number of unique values for each variable, the construct
  sapply(sapply(mf, unique), length)
        ^^^

only works if the number of unique values is _not_ the same for all
variables. An lapply() rather than an sapply() was intended here because
the outcome needs to be a _l_ist which is only the case if it cannot be
_s_implified.

Replaced now by

  sapply(mf, function(var) length(unique(var)))

which directly computes the number of unique values.



git-svn-id: svn://svn.r-forge.r-project.org/svnroot/partykit@1063 255f4249-2263-4417-b1ea-8498779e27a7